move tilemap class to new file
